Added back ray tracing code to shader, but with no shapes yet
